scripts/requirements.txt
pillow
fuzzywuzzy
python-Levenshtein
firebase-admin
